+++ lib/token.js
@@ -1,46 +1,13 @@
 /* eslint-disable complexity, max-statements */
 var http          = require('./http');
 var util          = require('./util');
+var oauthUtil     = require('./oauthUtil');
 var Q             = require('q');
 var sdkCrypto     = require('./crypto');
 var AuthSdkError  = require('./errors/AuthSdkError');
 var OAuthError    = require('./errors/OAuthError');
 var config        = require('./config');
 var cookies       = require('./cookies');
-
-function getWellKnown(sdk) {
-  return http.get(sdk, sdk.options.url + '/.well-known/openid-configuration');
-}
-
-function validateClaims(sdk, claims, aud, iss) {
-  if (!claims || !iss || !aud) {
-    throw new AuthSdkError('The jwt, iss, and aud arguments are all required');
-  }
-
-  var now = Math.floor(new Date().getTime()/1000);
-
-  if (claims.iss !== iss) {
-    throw new AuthSdkError('The issuer [' + claims.iss + '] ' +
-      'does not match [' + iss + ']');
-  }
-
-  if (claims.aud !== aud) {
-    throw new AuthSdkError('The audience [' + claims.aud + '] ' +
-      'does not match [' + aud + ']');
-  }
-
-  if (claims.iat > claims.exp) {
-    throw new AuthSdkError('The JWT expired before it was issued');
-  }
-
-  if ((now - sdk.options.maxClockSkew) > claims.exp) {
-    throw new AuthSdkError('The JWT expired and is no longer valid');
-  }
-
-  if (claims.iat > (now + sdk.options.maxClockSkew)) {
-    throw new AuthSdkError('The JWT was issued in the future');
-  }
-}
 
 function decodeToken(token) {
   var jwt = token.split('.');
@@ -94,7 +61,7 @@ function verifyIdToken(sdk, idToken, options) {
     }
   }
 
-  return getWellKnown(sdk)
+  return oauthUtil.getWellKnown(sdk)
     .then(function(res) {
       return http.get(sdk, res['jwks_uri']);
     })
@@ -132,37 +99,6 @@ function refreshIdToken(sdk, options) {
   return getToken(sdk, options);
 }
 
-function loadFrame(src) {
-  var iframe = document.createElement('iframe');
-  iframe.style.display = 'none';
-  iframe.src = src;
-
-  return document.body.appendChild(iframe);
-}
-
-function loadPopup(src, options) {
-  var title = options.popupTitle || 'External Identity Provider User Authentication';
-  var appearance = 'toolbar=no, scrollbars=yes, resizable=yes, ' +
-    'top=100, left=500, width=600, height=600';
-  return window.open(src, title, appearance);
-}
-
-function addListener(eventTarget, name, fn) {
-    if (eventTarget.addEventListener) {
-      eventTarget.addEventListener(name, fn);
-    } else {
-      eventTarget.attachEvent('on' + name, fn);
-    }
-  }
-
-function removeListener(eventTarget, name, fn) {
-  if (eventTarget.removeEventListener) {
-    eventTarget.removeEventListener(name, fn);
-  } else {
-    eventTarget.detachEvent('on' + name, fn);
-  }
-}
-
 function addPostMessageListener(sdk, timeout, state) {
   var deferred = Q.defer();
 
@@ -175,41 +111,12 @@ function addPostMessageListener(sdk, timeout, state) {
     deferred.resolve(e.data);
   }
 
-  addListener(window, 'message', responseHandler);
+  oauthUtil.addListener(window, 'message', responseHandler);
 
   return deferred.promise.timeout(timeout || 120000, new AuthSdkError('OAuth flow timed out'))
     .fin(function() {
-      removeListener(window, 'message', responseHandler);
+      oauthUtil.removeListener(window, 'message', responseHandler);
     });
-}
-
-function hashToObject(hash) {
-  // Predefine regexs for parsing hash
-  var plus2space = /\+/g;
-  var paramSplit = /([^&=]+)=?([^&]*)/g;
-
-  // Remove the leading hash
-  var fragment = hash.substring(1);
-
-  var obj = {};
-
-  // Loop until we have no more params
-  var param;
-  while (true) { // eslint-disable-line no-constant-condition
-    param = paramSplit.exec(fragment);
-    if (!param) { break; }
-
-    var key = param[1];
-    var value = param[2];
-
-    // id_token should remain base64url encoded
-    if (key === 'id_token' || key === 'access_token' || key === 'code') {
-      obj[key] = value;
-    } else {
-      obj[key] = decodeURIComponent(value.replace(plus2space, ' '));
-    }
-  }
-  return obj;
 }
 
 function addFragmentListener(sdk, windowEl, timeout) {
@@ -227,7 +134,7 @@ function addFragmentListener(sdk, windowEl, timeout) {
       if (windowEl &&
           windowEl.location &&
           windowEl.location.hash) {
-        deferred.resolve(hashToObject(windowEl.location.hash));
+        deferred.resolve(oauthUtil.hashToObject(windowEl.location.hash));
       } else if (windowEl && !windowEl.closed) {
         setTimeout(hashChangeHandler, 500);
       }
@@ -239,77 +146,6 @@ function addFragmentListener(sdk, windowEl, timeout) {
   hashChangeHandler();
 
   return deferred.promise.timeout(timeout || 120000, new AuthSdkError('OAuth flow timed out'));
-}
-
-function getOAuthUrls(sdk, oauthParams, options) {
-  options = options || {};
-
-  // Get user-supplied arguments
-  var authorizeUrl = util.removeTrailingSlash(options.authorizeUrl) || sdk.options.authorizeUrl;
-  var issuer = util.removeTrailingSlash(options.issuer) || sdk.options.issuer;
-  var userinfoUrl = util.removeTrailingSlash(options.userinfoUrl) || sdk.options.userinfoUrl;
-
-  // If an issuer exists, but it's not a url
-  if (issuer && !(/^https?:/.test(issuer))) {
-    // Make it a url
-    issuer = sdk.options.url + '/oauth2/' + issuer;
-  }
-
-  // If an authorizeUrl is supplied without an issuer, and an id_token is requested
-  if (authorizeUrl &&
-      !issuer &&
-      oauthParams.responseType.indexOf('id_token') !== -1) {
-    // The issuer is ambiguous, so we won't be able to validate the id_token jwt
-    throw new AuthSdkError('Cannot request idToken with an authorizeUrl without an issuer');
-  }
-
-  // If a token is requested without an issuer
-  if (!issuer && oauthParams.responseType.indexOf('token') !== -1) {
-    // If an authorizeUrl is supplied without a userinfoUrl
-    if (authorizeUrl && !userinfoUrl) {
-      // The userinfoUrl is ambiguous, so we won't be able to call getUserInfo
-      throw new AuthSdkError('Cannot request accessToken with an authorizeUrl without an issuer or userinfoUrl');
-    }
-
-    // If a userinfoUrl is supplied without a authorizeUrl
-    if (userinfoUrl && !authorizeUrl) {
-      // The authorizeUrl is ambiguous, so we won't be able to call the authorize endpoint
-      throw new AuthSdkError('Cannot request token with an userinfoUrl without an issuer or authorizeUrl');
-    }
-  }
-
-  var isSharedResourceServer = new RegExp('^https?://.*?/oauth2/.+');
-
-  // If issuer is supplied, but not authorizeUrl
-  if (issuer && !authorizeUrl) {
-    // Build authorizeUrl until we create it dynamically
-    if (isSharedResourceServer.test(issuer)) {
-      authorizeUrl = issuer + '/v1/authorize';
-    } else {
-      authorizeUrl = issuer + '/oauth2/v1/authorize';
-    }
-  }
-
-  // If issuer is supplied, but not userinfoUrl
-  if (issuer && !userinfoUrl) {
-    // Build userinfoUrl until we create it dynamically
-    if (isSharedResourceServer.test(issuer)) {
-      userinfoUrl = issuer + '/v1/userinfo';
-    } else {
-      userinfoUrl = issuer + '/oauth2/v1/userinfo';
-    }
-  }
-
-  // Anything not set should be defaulted
-  issuer = issuer || sdk.options.url;
-  authorizeUrl = authorizeUrl || issuer + '/oauth2/v1/authorize';
-  userinfoUrl = userinfoUrl || issuer + '/oauth2/v1/userinfo';
-
-  return {
-    issuer: issuer,
-    authorizeUrl: authorizeUrl,
-    userinfoUrl: userinfoUrl
-  };
 }
 
 function handleOAuthResponse(sdk, oauthParams, res, urls) {
@@ -334,7 +170,7 @@ function handleOAuthResponse(sdk, oauthParams, res, urls) {
     }
 
     var clientId = oauthParams.clientId || sdk.options.clientId;
-    validateClaims(sdk, jwt.payload, clientId, urls.issuer);
+    oauthUtil.validateClaims(sdk, jwt.payload, clientId, urls.issuer);
 
     var idToken = {
       idToken: res['id_token'],
@@ -543,7 +379,7 @@ function getToken(sdk, oauthOptions, options) {
       urls;
   try {
     // Get authorizeUrl and issuer
-    urls = getOAuthUrls(sdk, oauthParams, options);
+    urls = oauthUtil.getOAuthUrls(sdk, oauthParams, options);
     requestUrl = urls.authorizeUrl + buildAuthorizeParams(oauthParams);
   } catch (e) {
     return Q.reject(e);
@@ -568,7 +404,7 @@ function getToken(sdk, oauthOptions, options) {
   switch (flowType) {
     case 'IFRAME':
       var iframePromise = addPostMessageListener(sdk, options.timeout, oauthParams.state);
-      var iframeEl = loadFrame(requestUrl);
+      var iframeEl = oauthUtil.loadFrame(requestUrl);
       return iframePromise
         .then(function(res) {
           return handleOAuthResponse(sdk, oauthParams, res, urls);
@@ -595,7 +431,7 @@ function getToken(sdk, oauthOptions, options) {
       var windowOptions = {
         popupTitle: options.popupTitle
       };
-      var windowEl = loadPopup(requestUrl, windowOptions);
+      var windowEl = oauthUtil.loadPopup(requestUrl, windowOptions);
 
       // Poll until we get a valid hash fragment
       if (oauthParams.responseMode === 'fragment') {
@@ -683,7 +519,7 @@ function getWithRedirect(sdk, oauthOptions, options) {
     }
   }
   
-  var urls = getOAuthUrls(sdk, oauthParams, options);
+  var urls = oauthUtil.getOAuthUrls(sdk, oauthParams, options);
   var requestUrl = urls.authorizeUrl + buildAuthorizeParams(oauthParams);
 
   // Set session cookie to store the oauthParams
@@ -698,17 +534,8 @@ function getWithRedirect(sdk, oauthOptions, options) {
   sdk.token.getWithRedirect._setLocation(requestUrl);
 }
 
-function isToken(obj) {
-  if (obj &&
-      (obj.accessToken || obj.idToken) &&
-      Array.isArray(obj.scopes)) {
-    return true;
-  }
-  return false;
-}
-
 function refreshToken(sdk, token) {
-  if (!isToken(token)) {
+  if (!oauthUtil.isToken(token)) {
     return Q.reject(new AuthSdkError('Refresh must be passed a token with ' +
       'an array of scopes and an accessToken or idToken'));
   }
@@ -749,7 +576,7 @@ function parseFromUrl(sdk, url) {
       config.REDIRECT_OAUTH_PARAMS_COOKIE_NAME + ' cookie: ' + e.message));
   }
 
-  return Q.resolve(hashToObject(hash))
+  return Q.resolve(oauthUtil.hashToObject(hash))
     .then(function(res) {
       return handleOAuthResponse(sdk, oauthParams, res, urls);
     });
@@ -757,7 +584,7 @@ function parseFromUrl(sdk, url) {
 
 function getUserInfo(sdk, accessTokenObject) {
   if (!accessTokenObject ||
-      (!isToken(accessTokenObject) && !accessTokenObject.accessToken && !accessTokenObject.userinfoUrl)) {
+      (!oauthUtil.isToken(accessTokenObject) && !accessTokenObject.accessToken && !accessTokenObject.userinfoUrl)) {
     return Q.reject(new AuthSdkError('getUserInfo requires an access token object'));
   }
   return http.httpRequest(sdk, {
